requirements.txt
# Web Scraping
playwright==1.49.0
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3

# Database
sqlite3  # built-in

# AI/LLM
google-generativeai==0.8.3

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Utilities
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1

# HuggingFace
huggingface-hub==0.27.0
